browser-agent/pyproject.toml
[project]
name = "browser-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.70.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
